Abstracte the LogitsDecoder
